# api/bff-openapi.yaml
...
openapi: "3.0.3"
info:
  title: Skateboard UI-backend API
  version: "1.0.0"
...
paths:
  /api/podcast:
    get:
      tags: [podcast]
      operationId: getPodcastFeed
      summary: Get published podcast posts (paginated)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_TAB_PODCAST
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 10 }
        - name: search
          in: query
          description: Case-insensitive title filter (used by the Home Featured Player episode picker)
          schema: { type: string }
      responses:
        '200':
          description: Podcast feed page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedPageResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [podcast]
      operationId: createPodcastPost
      summary: Create a new podcast post (admin only)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_PODCAST_CREATE_POST
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Podcast post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/podcast/{slug}:
    get:
      tags: [podcast]
      operationId: getPodcastPostBySlug
      summary: Get a single podcast post by slug
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_TAB_PODCAST
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Podcast post detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/podcast/{id}:
    put:
      tags: [podcast]
      operationId: updatePodcastPost
      summary: Update an existing podcast post (admin only)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_PODCAST_EDIT_POST
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
      responses:
        '200':
          description: Podcast post updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [podcast]
      operationId: deletePodcastPost
      summary: Delete a podcast post (admin only)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_PODCAST_DELETE_POST
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Podcast post deleted
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/podcast/import:
    post:
      tags: [podcast]
      operationId: importPodcastPosts
      summary: Bulk import podcast posts from JSON (admin only)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_PODCAST_IMPORT_JSON
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportPostsRequest'
      responses:
        '200':
          description: Import result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResult'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/podcast/sync:
    post:
      tags: [podcast]
      operationId: syncPodcastFromYoutube
      summary: Trigger the same YouTube sync the scheduler runs, on demand (admin only)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_PODCAST_IMPORT_JSON
      responses:
        '200':
          description: Sync result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncResultResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/categories:
    get:
      tags: [podcast]
      operationId: getCategories
      summary: List enabled categories (mirrors the channel's YouTube playlists)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_TAB_PODCAST
      responses:
        '200':
          description: Enabled categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/categories/{slug}/posts:
    get:
      tags: [podcast]
      operationId: getCategoryPosts
      summary: Get published posts for a category (paginated)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_TAB_PODCAST
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: Category feed page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedPageResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/categories:
    get:
      tags: [podcast]
      operationId: getAdminCategories
      summary: List every category, disabled ones included (admin only)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_PODCAST_MANAGE_CATEGORIES
      responses:
        '200':
          description: All categories in display order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminCategoryResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/categories/{id}:
    patch:
      tags: [podcast]
      operationId: updateCategory
      summary: Rename a category (admin only); null name resets to the YouTube title
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_PODCAST_MANAGE_CATEGORIES
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'
      responses:
        '200':
          description: Updated category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCategoryResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/categories/order:
    put:
      tags: [podcast]
      operationId: reorderCategories
      summary: Set the display order of all categories (admin only)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_PODCAST_MANAGE_CATEGORIES
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderCategoriesRequest'
      responses:
        '200':
          description: Categories in their new order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminCategoryResponse'
        '400':
          description: Unknown, duplicated, or missing category ids
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/categories/{id}/default:
    put:
      tags: [podcast]
      operationId: setDefaultCategory
      summary: Make this category the single default (admin only)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_PODCAST_MANAGE_CATEGORIES
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: The new default category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCategoryResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SocialMediaLink:
      type: object
      required: [url]
      properties:
        platform:
          type: string
        url:
          type: string
    PostResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [draft, scheduled, published]
        publishAt:
          type: string
          format: date-time
        coverUrl:
          type: string
        blocks:
          type: array
          items:
            type: object
            additionalProperties: true
          description: Ordered array of typed content blocks
        socialMediaLinks:
          type: array
          items:
            $ref: '#/components/schemas/SocialMediaLink'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        youtubeVideoId:
          type: string
          nullable: true
          description: Set only for posts ingested from the YouTube sync job.
        youtubeUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
          description: Plain-text video description from YouTube; not present on manually-authored posts predating the sync.
        durationSeconds:
          type: integer
          nullable: true
        episodeNumber:
          type: integer
          nullable: true
    FeedPageResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: '#/components/schemas/PostResponse'
        total:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer
    CreatePostRequest:
      type: object
      required: [title, coverUrl]
      properties:
        title:
          type: string
        coverUrl:
          type: string
        status:
          type: string
          enum: [draft, scheduled, published]
          default: published
        publishAt:
          type: string
          format: date-time
        blocks:
          type: array
          items:
            type: object
            additionalProperties: true
        socialMediaLinks:
          type: array
          items:
            $ref: '#/components/schemas/SocialMediaLink'
    UpdatePostRequest:
      type: object
      required: [title, coverUrl]
      properties:
        title:
          type: string
        coverUrl:
          type: string
        status:
          type: string
          enum: [draft, scheduled, published]
        publishAt:
          type: string
          format: date-time
        blocks:
          type: array
          items:
            type: object
            additionalProperties: true
        socialMediaLinks:
          type: array
          items:
            $ref: '#/components/schemas/SocialMediaLink'
    ImportPostsRequest:
      type: object
      required: [posts]
      properties:
        posts:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ImportPostItem'
    ImportPostItem:
      type: object
      required: [title]
      properties:
        title:
          type: string
        coverUrl:
          type: string
        status:
          type: string
          enum: [draft, scheduled, published]
          default: published
        publishAt:
          type: string
          format: date-time
        blocks:
          type: array
          items:
            type: object
            additionalProperties: true
        socialMediaLinks:
          type: array
          items:
            $ref: '#/components/schemas/SocialMediaLink'
    ImportResult:
      type: object
      properties:
        imported:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            type: string
    CategoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        coverUrl:
          type: string
          nullable: true
        default:
          type: boolean
        postCount:
          type: integer
          format: int64
    AdminCategoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
          description: Effective display name (customName when set, else the YouTube title).
        youtubeName:
          type: string
          description: The playlist title as the sync last saw it.
        customName:
          type: string
          nullable: true
          description: Admin rename override; null when the YouTube title is used as-is.
        coverUrl:
          type: string
          nullable: true
        enabled:
          type: boolean
          description: False when the source playlist disappeared from YouTube.
        default:
          type: boolean
        displayOrder:
          type: integer
          nullable: true
          description: Admin-set position; null sorts after all positioned categories.
        postCount:
          type: integer
          format: int64
    UpdateCategoryRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
          description: New display name; null or blank resets to the YouTube title.
    ReorderCategoriesRequest:
      type: object
      required: [categoryIds]
      properties:
        categoryIds:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
    SyncResultResponse:
      type: object
      properties:
        received:
          type: integer
        created:
          type: integer
        existing:
          type: integer
        categoryChanges:
          type: integer
        success:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time